chore: create example package
